requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
openai==0.28.1
pinecone-client==3.0.0
python-multipart==0.0.6
httpx==0.23.3
numpy==1.24.3
tiktoken==0.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0 